+++ src/components/Carrinho.js
@@ -1,17 +1,73 @@
-import React from "react";
+import React from 'react'
 import { StyledContainer, ItemCarrinho } from "./styled";
 
-export function Carrinho(props) {
-  return (
-    <StyledContainer>
-      <h2>Carrinho: </h2>
-      {props.carrinho.map((produto) => {
+class Carrinho extends React.Component {
+    state={
+        quantidade: 4,
+        listaDeCompras: [{
+            produto:"Item A",
+            preco: 5,
+            key: 150
+        },
+        {
+            produto:"Item B",
+            preco: 10,
+            key: 160
+        },
+        {
+            produto:"Item E",
+            preco: 5,
+            key: 170
+        },
+        {
+            produto:"Item D",
+            preco: 10,
+            key: 180
+        },
+        {
+            produto:"Item C",
+            preco: 10,
+            key: 190
+        }],
+
+    }
+
+    onClickElimina = (chave) => {
+
+       let novaLista = this.state.listaDeCompras.filter((item) => {
+        return chave !== item.key
+        })
+
+        this.setState({listaDeCompras: novaLista})
+    }
+
+    render() {
+        let precoTotal = 0
+
+            this.state.listaDeCompras.map((item) => {
+                 precoTotal += item.preco
+            })
+
+
+            let produto = this.state.listaDeCompras.map((item) => {
+                return (
+                    <ItemCarrinho>
+                        <p>{this.state.quantidade} x {item.produto}</p>
+                        <p onClick={() => this.onClickElimina(item.key)} className="elimina">X</p>
+                    </ItemCarrinho>
+                )
+           })
+
+
         return (
-          <ItemCarrinho>
-            <p>{produto.nome} {produto.valor}</p>
-          </ItemCarrinho>
-        );
-      })}
-    </StyledContainer>
-  );
+            <StyledContainer>
+                <h3>Carrinho:</h3>
+                {produto}
+                <p>Total: R${precoTotal}</p>
+            </StyledContainer>
+        )
+    }
+
 }
+
+export default Carrinho
